requirements.txt
Pillow
ffmpeg-python
tkinterdnd2
tk